Create fb_ads.py
The project concerns how facebook ad targeting and potential reach works. There is a predefined users on facebook(in this experimental case). When the user selects a state or states, the values in those states are summed up to be his available audience by location. When you user selects from the interest targeting, the available audience is narrowed down, based on the interest selected.